Use tracetest instead of oteltest
