+++ .github/dependabot.yml
@@ -197,6 +197,16 @@ updates:
       day: "sunday"
   -
     package-ecosystem: "gomod"
+    directory: "/instrumentation/github.com/emicklei/go-restful/otelrestful/test"
+    labels:
+      - dependencies
+      - go
+      - "Skip Changelog"
+    schedule:
+      interval: "weekly"
+      day: "sunday"
+  -
+    package-ecosystem: "gomod"
     directory: "/instrumentation/github.com/gin-gonic/gin/otelgin"
     labels:
       - dependencies
