+++ instrumentation/github.com/emicklei/go-restful/otelrestful/go.mod
@@ -14,6 +14,5 @@ require (
 	go.opentelemetry.io/contrib v0.22.0
 	go.opentelemetry.io/contrib/propagators/b3 v0.22.0
 	go.opentelemetry.io/otel v1.0.0-RC2
-	go.opentelemetry.io/otel/oteltest v1.0.0-RC2
 	go.opentelemetry.io/otel/trace v1.0.0-RC2
 )
